api common for worker and pool
